@@ -9,41 +9,42 @@
   },
   "dependencies": {
     "@fiahfy/electron-accelerator-formatter": "^0.0.4",
-    "@fiahfy/electron-context-menu": "^0.0.5",
+    "@fiahfy/electron-context-menu": "^0.0.6",
     "@fiahfy/simple-clone": "^0.0.3",
-    "@mdi/font": "^4.9.95",
-    "@nuxtjs/vuetify": "^1.11.0",
-    "@types/debounce": "^1.2.0",
-    "@types/node": "^12.0.12",
-    "d3": "^5.15.0",
+    "@mdi/font": "^5.3.45",
+    "@nuxtjs/vuetify": "^1.11.2",
+    "@vue/composition-api": "^0.6.6",
+    "d3": "^5.16.0",
     "debounce": "^1.2.0",
-    "electron-devtools-installer": "^2.2.4",
+    "electron-devtools-installer": "^3.0.0",
     "electron-window-state": "^5.0.3",
-    "nuxt": "^2.11.0",
+    "nuxt": "^2.13.0",
     "pretty-bytes": "^5.3.0",
     "typeface-roboto": "^0.0.75",
-    "vue-property-decorator": "^8.4.0",
-    "vuex-module-decorators": "^0.16.1",
-    "vuex-persistedstate": "^2.7.1"
+    "vue-property-decorator": "^8.4.2",
+    "vuex-module-decorators": "^0.17.0",
+    "vuex-persistedstate": "^3.0.1"
   },
   "devDependencies": {
     "@fiahfy/icns-convert": "^0.0.8",
     "@fiahfy/ico-convert": "^0.0.7",
-    "@nuxt/typescript-build": "^0.5.8",
-    "@nuxtjs/eslint-config-typescript": "^1.0.2",
+    "@nuxt/typescript-build": "^1.0.3",
+    "@nuxtjs/eslint-config-typescript": "^2.0.0",
     "@types/d3": "^5.7.2",
-    "cross-env": "^7.0.0",
-    "electron": "^8.0.2",
-    "electron-builder": "^22.3.2",
-    "eslint": "^6.8.0",
-    "eslint-config-prettier": "^6.10.0",
-    "eslint-plugin-prettier": "^3.1.2",
-    "husky": "^4.2.3",
-    "node-sass": "^4.13.1",
+    "@types/debounce": "^1.2.0",
+    "@types/node": "^12.12.47",
+    "cross-env": "^7.0.2",
+    "electron": "^8.3.3",
+    "electron-builder": "^22.7.0",
+    "eslint": "^7.3.0",
+    "eslint-config-prettier": "^6.11.0",
+    "eslint-plugin-prettier": "^3.1.4",
+    "husky": "^4.2.5",
+    "node-sass": "^4.14.1",
     "npm-run-all": "^4.1.5",
-    "prettier": "^1.19.1",
+    "prettier": "^2.0.5",
     "sass-loader": "^8.0.2",
-    "vuetify-loader": "^1.4.3",
+    "vuetify-loader": "^1.5.0",
     "worker-loader": "^2.0.0"
   },
   "keywords": [
