@@ -1,4 +1,4 @@
-const pkg = require('./package')
+import pkg from './package'
 
 module.exports = {
   /*
@@ -9,15 +9,15 @@ module.exports = {
       // Extend only webpack config for client-bundle
       if (ctx.isClient) {
         config.target = 'electron-renderer'
-        // config.output.globalObject = 'this'
-        // config.module.rules.unshift({
-        //   test: /\.worker\.ts$/,
-        //   loader: 'worker-loader'
-        // })
+        config.output.globalObject = 'this'
+        config.module.rules.unshift({
+          test: /\.worker\.ts$/,
+          loader: 'worker-loader',
+        })
       }
       // Set relative path
       config.output.publicPath = './_nuxt/'
-    }
+    },
   },
 
   /*
@@ -28,13 +28,17 @@ module.exports = {
   /*
    ** Global CSS
    */
-  css: ['@mdi/font/css/materialdesignicons.css', 'typeface-roboto/index.css'],
+  css: [
+    '@mdi/font/css/materialdesignicons.css',
+    'typeface-roboto/index.css',
+    '~/assets/app.scss',
+  ],
 
   /*
    ** Generate configuration
    */
   generate: {
-    dir: 'app'
+    dir: 'app',
   },
 
   /*
@@ -47,7 +51,7 @@ module.exports = {
    */
   head: {
     title: pkg.productName,
-    meta: [{ hid: 'charset', charset: 'utf-8' }]
+    meta: [{ hid: 'charset', charset: 'utf-8' }],
   },
 
   /*
@@ -69,17 +73,17 @@ module.exports = {
             light: {
               primary: '#ff4081',
               secondary: '#424242',
-              accent: '#ff4081'
+              accent: '#ff4081',
             },
             dark: {
               primary: '#ff4081',
               secondary: '#E0E0E0',
-              accent: '#ff4081'
-            }
-          }
-        }
-      }
-    ]
+              accent: '#ff4081',
+            },
+          },
+        },
+      },
+    ],
   ],
 
   /*
@@ -88,16 +92,18 @@ module.exports = {
   plugins: [
     '~/plugins/electron-accelerator-formatter',
     '~/plugins/electron-context-menu',
+    '~/plugins/event-bus',
     '~/plugins/ipc',
     '~/plugins/pretty-bytes',
-    '~/plugins/vuex-persistedstate'
+    '~/plugins/vue-composition-api',
+    '~/plugins/vuex-persistedstate',
   ],
 
   /*
    ** Router configuration
    */
   router: {
-    mode: 'hash'
+    mode: 'hash',
   },
 
   /*
@@ -110,7 +116,7 @@ module.exports = {
    */
   vue: {
     config: {
-      productionTip: false
-    }
-  }
+      productionTip: false,
+    },
+  },
 }
