@@ -3,11 +3,6 @@ module.exports = {
   extends: [
     '@nuxtjs/eslint-config-typescript',
     'plugin:prettier/recommended',
-    'prettier/vue'
+    'prettier/vue',
   ],
-  rules: {
-    // @see: https://github.com/typescript-eslint/typescript-eslint/issues/1220
-    'no-unused-expressions': 'off',
-    '@typescript-eslint/no-unused-expressions': 'error'
-  }
 }
