@@ -1,10 +1,9 @@
 import { Plugin } from '@nuxt/types'
 import createPersistedState from 'vuex-persistedstate'
+import '~/store' // For reloading if store is modified
 
 const persistedStatePlugin: Plugin = ({ store }) => {
-  ;(<any>window).onNuxtReady(() => {
-    createPersistedState()(store)
-  })
+  createPersistedState()(store)
 }
 
 export default persistedStatePlugin
